refactor: 배포를 위한 pages 파일 개선 및 콘솔 삭제
