+++ components/Effect/Effect.tsx
@@ -1,10 +1,7 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { EffectContainer } from './style';
 
 const Effect = ({ effect, medicines }) => {
-  useEffect(() => {
-    console.log('effect', effect, medicines);
-  }, [effect, medicines]);
   return (
     <EffectContainer>
       <div>{effect}</div>
