+++ pages/input/index.tsx
@@ -1,4 +1,3 @@
-import { useRouter } from 'next/router';
 import { useState } from 'react';
 import { Button } from '@mui/material';
 import InfoModal from '../../components/InfoModal/InfoModal';
@@ -6,11 +5,21 @@ import { useSelector } from 'react-redux';
 import { selectPrescription } from '../../store/slices/prescription';
 import { getInputInfo, patchInputInfo } from '../../apis/info';
 import { getPrescriptionInputValue } from '../../utils/input';
-import {
-  PrescriptionContainer,
-  PrescriptionInfo,
-  PrescriptionInput,
-} from './style';
+import styled from '@emotion/styled';
+export const PrescriptionContainer = styled.div`
+  display: flex;
+  flex-direction: row;
+  height: 3rem;
+`;
+
+export const PrescriptionInfo = styled.div`
+  margin: 10px;
+`;
+
+export const PrescriptionInput = styled.input`
+  margin: 5px;
+  width: 3rem;
+`;
 
 const Input = () => {
   const [isModalOpen, setIsModalOpen] = useState(true);
