--- pages/input/style.ts
@@ -1,16 +1,0 @@
-import styled from '@emotion/styled';
-
-export const PrescriptionContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  height: 3rem;
-`;
-
-export const PrescriptionInfo = styled.div`
-  margin: 10px;
-`;
-
-export const PrescriptionInput = styled.input`
-  margin: 5px;
-  width: 3rem;
-`;
