--- pages/login/style.ts
@@ -1,13 +1,0 @@
-import styled from '@emotion/styled';
-
-export const LoginContainer = styled.div`
-  max-width: 640px;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 1rem;
-  margin: 0 auto;
-  padding: 15rem 0;
-`;
