--- pages/signup/style.ts
@@ -1,30 +1,0 @@
-import styled from '@emotion/styled';
-
-export const FormContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const StyledForm = styled.form`
-  width: 392px;
-  height: 444px;
-  padding: 1rem 1rem 0;
-`;
-
-export const FormInputContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-top: 2.5rem;
-  gap: 2rem;
-`;
-
-export const MbtiContainer = styled.div`
-  display: flex;
-  gap: 1rem;
-`;
-
-export const MbtiWrapper = styled.div`
-  width: calc(50%);
-`;
