+++ utils/crud.ts
@@ -56,3 +56,9 @@ export const deleteById = async (id: number) => {
   const response = await apiClient.delete(`/${id}`);
   return response.data;
 };
+
+export const queryKeys = {
+  posts: ['posts'] as const,
+  createPost: (id: number, title: string, body: number) =>
+    ['createPost', id, title, body] as const,
+};
